e-commerce-chatbot-end_to_end/requirements.txt
streamlit 

google-generativeai

langchain-google-genai

python-dotenv 

flask 

PyPDF2

langchain 

langchain-astradb 

